use clang-tidy command with task to generate items in problem's view
Enable running the `clang-tidy` C/C++ linting present from clang-tidy as a task.
With the new updates, linting can be performed across an entire workspace using user
specified checks and linting rules. The problem markers collected when executing the
task will be recorded and displayed in the `problems-widget`.
The result in the "Problems" view is persistent until you
re-generates the task.

Signed-off-by: Jacques Bouthillier <jacques.bouthillier@ericsson.com>