[security] Bump lodash.template to 4.5.0
There is a "high level" security advisory about lodash.template <4.5.0.

lodash.template is not a direct dependency for us. As of current theia master
it's pulled indirectly, as a dev-dependency, from 2 paths:

- v3.6.2: Hoisted from "_project_#remap-istanbul#gulp-util#lodash.template"
- v4.4.0: Hoisted from "_project_#lerna#conventional-recommended-bump#git-raw-commits"
          Hoisted from "_project_#lerna#conventional-changelog-cli#conventional-changelog#conventional-changelog-core#git-raw-commits"

For the first case, updating "remap-istanbul" to latest, in our project's main
package.json, removes the indirect dependency to lodash.template.

For the second case, we need to update the chain from the beginning, which
means using a new major lerna version (3.x), so that at the end of the chain,
git-raw-commits will be ok to pull lodash.template 4.5.0.

This is a WIP, attempting to do this. I am concerned that the major lerna version
change may have unintended consequences.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>