core: catch command handler errors
If a contribution throws an error during basic lifecycle steps, it might
kill whatever is trying to iterate over all the contributions.

A better strategy would be to catch errorred handlers and keep looking
for other contributions.

Also fix `.getToggledHandler` to not stop prematurely and actually test
if the handler is toggled or not.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>