[Eclipse] update references to old GH org
Update all URLs pointing to the old organization so that they now
refer to "eclipe-theia". GitHub magically redirects but I am not
certain if we can assume this will work forever, so I even updated
links that are part of the code, e.g. that point to an issue or
GH comment.

Fixes #6169

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>