fix problem matcher items in the prompt before task is started
When user starts a task that is not associated with any problem
matchers, s/he should be prompted with a list of problem matchers that
are already defined in Theia. In the current version of Theia, the
problem matcher list is broken by 657bda85aeedcc91cd295d4a4eba967ed0fb79c3, and this change fixes the problem.

Signed-off-by: Liang Huang <liang.huang@ericsson.com>