Adjust tree node foreground color when active
Fixes #3960
Fixes #6831

With this change, when a tree node is active, both the content including the node itself,
the `tail` and `suffixes` are displayed with the same `activeSelectionForeground` which helps in
visibility, the overall look, and also aligns with VS Code. Previously, for
certain themes, the colors were difficult to see.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>