[debug] Lazily update frames of all threads in all-stop mode
When a breakpoint occurs in a thread in all-stop mode, all other
threads have outdated/missing stackframes which means no variables
show up and can be inspected. This fixes the issue by marking each
thread as potentially having "dirty" frames: frames that need to be
re-fetched when the user selects them.

Signed-off-by: Samuel Frylmark <samuel.frylmark@ericsson.com>