Navigator:Select Folder when user creates one.
Problem Before: When you create a new folder, the folder is not selected. However, in VSCode it is.
Solution: Updated the Workspace service to contain the createFile and createFolder methods, which would call the filesystem to create it for us, then an event is fired so that the navigator-widget would know about it, once the file receives the uri. It will then create a URI, get the nodes by that URI, and then select that specific node.

Issue ID: 6190

Signed-off-by: Muhammad Anas Shahid <muhammad.shahid@ericsson.com>