Core: Do no expand the widgets on the side-bars for the context menu
Fixes: https://github.com/eclipse-theia/theia/issues/4367

Before, right clicking on different menus would focus the menu item and open it, however this should not be the case. There was a dangling code in the handleContextMenu which causes this effect, as it was checking for the id when right clicked and looked up the ID for the menu-item and set the current title to the menu-item.

Signed-off-by: Muhammad Anas Shahid <muhammad.shahid@mail.mcgill.ca>