getting-started: use absolute link for logo
This commit updates the `getting-started` readme to use
the external absolute link for the theia extension logo.
This change is useful when embedding the document in other
pages (ex: typedoc rendering).

A similar change was done in the past (https://github.com/eclipse-theia/theia/pull/4790),
but it requires the logo to exist in the repo. Now that the logo exists, we can now
update the link so it is absolute.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>