Fixes: #7170 , #5027
Title: [editor] Show languages while preview

This issue fixes the underlying problem where when a new file was opened
in preview mode, the focus was on the previously active file and not on
the navigator + the statusbar at the bottom was not updated. This PR
collectively fixes both the issues.

Signed-off-by: Muhammad Anas Shahid <muhammad.shahid@ericsson.com>