markers: update format of markers
This commit updates the formatting of markers present in the `problems-view`
to be consistent with VS Code. Namely, the commit enforces that the marker
message is the most important data, that the `owner` or `source` is displayed with the
`code` if present, and the `line number` and `column number` are displayed last.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>