[example apps] update html-language-features vscode built-in
We discovered that this built-in extension requires to have
typescript bundled, to handle HTML-embedded JS/TS.

This PR updates the vscode built-ins so that a fixed version
of that extension is used. I updated the html built-in too,
just in case.

See upstream issue: https://github.com/theia-ide/vscode-builtin-extensions/issues/33

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>