console: fix restoration of the console widget
The following pull-request fixes the restoration of the console widget
which previously threw an error when attempting to restore history to
the input widget as it was not yet created (async). The fix updates the
`restore` logic to wait until the input is created in order to restore
it's content.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>