[ip] pin jschardet to v1.6.0
Our jschardet has been stepped to a new major version and we missed
it. Since this dependency is licensed under the LGPL, it requires a
special premission from the Eclipse Board, before we are allowed
to use a new version, as part of a release.

There is only a bit more than 2 weeks before our next release, and
I do not know when the next board meeting may happen. So it looks
like we might have to delay the relase or momentarily go back to
the previous, approved, dependency version.

This is what this PR does, using a resolution.

CQ (need to be committer to see):
https://dev.eclipse.org/ipzilla/show_bug.cgi?id=22481

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>