navigator: handle empty multi-root workspace
When a multi-root workspace is opened, the navigator will try to draw
the file tree because `model.root` is a `WorkspaceNode` with no
children.

This commit handles this case by adding a bit of ui to help the user add
folders to his workspace, instead of displaying a blank view.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>