eslint: add 'no-tabs' rule
The following commit adds the `no-tabs` eslint rule to
enforce that `spaces` should always be used as according to the
project rules, and `editorconfig` properties.

The rule will now display errors for use of `tabs` over `spaces`,
and it ensures that new pull-requests do not break this pattern, and
cause unecessary changes when formatting the documents or performing a
`save`.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>