download-plugins: resolve plugin URLs on open-vsx
Extend plugins syntax from package.json to support both an URL (like before)
or to support a "version spec" following the form:

    publisher/extension@version

which will lookup the extension on https://open-vsx.org and download the file
from corresponding version.

Version can be anything that the `/api/{namespace}/{extension}/{version}` API
endpoint of open-vsx supports, ie. it can be an exact version number or `latest`.

This will use the plugin lock file to record the resolved URL, so that subsequent
downloads resolve to the same download URL when the plugin lock file is present.

Signed-off-by: Jérome Perrin <perrinjerome@gmail.com>