core: gracefully kill process tree on exit
When killing a given process, its children are supposed to keep running
unaffected. Since we never kill our subprocesses ourselves, we often end
up leaking orphaned processes. The most notorious is the plugin-host
process.

When running the browser version, yarn somehow ends up cleaning up for
us. But when running Electron, simply killing the backend process is not
enough, we need to find a way to kill everything.

This commit addresses this issue, with some platform specifities:

- On Windows we'll run `taskkill /f /t /pid xxx` in order to destroy
  the whole tree from the backend pid.
- On UNIX we'll prefer starting the backend process as group leader,
  this way we can kill the whole group at once. But we'll still also
  follow parent/child relationships in case some processes are spawned
  in a different group.

Signed-off-by: Paul <paul.marechal@ericsson.com>