core: lazy preference proxy
It is possible to register a preference schema to the
`PreferenceSchemaProvider` after the initial inversify binding phase.
This is problematic when wanting to do that late schema registration and
at the same time trying to bind a preference proxy because proxies must
be bound early on.

This commit fixes the issue by allowing `createPreferenceProxy` to
process a promise for its `schema` parameter (now `promisedSchema`) and
still return a proxy synchronously. Until the promise resolves, most
operations on the proxy will be no-ops.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>