mini-browser: fix host pattern logic
When we set environment variables in our launch.json file, we use a
notation like "VAR": "${env:VAR}". VS Code will replace the template in
the value string with the content of the actual env var value, and if
not set it will expand to "". This caused issues with the logic that
uses this value since we expected to get undefined instead of an empty
string.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>