getting-started: fix href links
The following commit adjusts the links in the `getting-started` view to
remove uses of `href="#"` which would modify the URL hash when the
commands are clicked. For the sake of the view, the href are not useful
as we want the links to be treated similarly to buttons.

The changes include:
- removing uses of `href="#"`.
- adjusting the CSS in `getting-started` to adjust the cursor for links.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>