vsx: update compatibility check for builtins
The changeset updates the logic when getting the latest compatible
version of a builtin extension (individual or pack) by fetching the
version which is defined by api version supported by the framework.

Since the `engines.vscode` property is generally not maintained by
vscode builtin extensions (since they are bundled with vscode at a
specific version) we cannot rely on the `engines.vscode` version to
fetch a compatible extension. Instead, we now fetch a version which
corresponds to the api level the framework defines.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>