Don't re-open hidden tree-view after reload
The PluginViewWidget has a property to suppress visibility updates
from a `when` clause in the case a user specifically requests to hide
it. However when the browser got reloaded, the flag was reset to
default and did not keep the hide request from the user.

This change preserves the flag mentioned above in the PluginViewWidget
state, this state is preserved and then available for use in
`restoreState` where the flag is now updated to the state before the
browser reloads.

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>