add EntryPointRegistry
Declared entry points in Theia extensions have their name prefixed with
the originating package name e.g. if the extension `@theia/a`
contributes an entry point named `b`, the name will be `@theia/a/b`.

Components that wish to use those entry points can query the
`EntryPointRegistry` to get the path on disk to the script.