Update CSS for description elements of tree nodes
Adds a new `theia-TreeNodeInfo` CSS class that should be used by every part of a treenode that adds additional info. The info part is displayed with a lower opacity.
Examples for this CSS class can be seen for the `owner` and `position` info of problem markers and the `path` of a debugging breakpoint.