application-package: support for yarn aliases (#9890)
This PR fixes #9856. It is done by changing the application package extension to use as a name for the packages (and more specifically theia extensions) the name used in the list of dependencies in `package.json` in place of the package name. In most cases these two names should match, but when yarn aliases are used they differ.

Signed-off-by: Federico Bozzini <federico.bozzini@gmail.com>
Co-authored-by: Paul Marechal <paul.marechal@ericsson.com>