core: fix `quick-commands`
The commit fixes the `quick-command` behavior which would not display
the list of recently used commands properly when users would use the
`quick-command` + type the prefix value, instead of triggering the
command directly. The change refactors the code to perform the `reset`
when we trigger the picks so we pick up the latest commands available.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>