Explicitly mark UnsafeWidgetUtilities
Create a new namespace for non-error-checked versions of Phosphor
utilities, replaces existing code blocks, and applies the new
utilities to cases that were throwing errors.

As the name implies, these utilities are unsafe because they
violate the expectations of Phosphor widgets. We should investigate
why we ever wrote code to do this in the first place, and fix it
if possible.

Signed-off-by: Colin Grant <colin.grant@ericsson.com>