cli: introduce exclude id list for extension-packs (#9956)
The commit introduces the ability to declare the list of plugin ids which should be explicitly excluded when we perform a download. The list of excluded plugin ids are used when we attempt to resolve plugins declared in extension-packs (extension-packs refer to plugins they wish to pull by id) but we want to exclude some problematic or unwanted plugins. This gives users the flexibility to consume extension-packs as builtins, but also be able to explicitly exclude ids they do not wish to pull as part of their application.

The `download:plugins` script was also refactored in order to make things clearer, more robust with better performance:
- extension-pack plugins are resolved and downloaded in parallel.
- if a pack is previously downloaded (and moved under the download folder at `.packs`) we will not re-download it or resolve it.
- the framework will not attempt to re-resolve packs at runtime, meaning excluded plugins will not re-appear.
- improved logging of the overall download and operations.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>
Co-authored-by: Paul Marechal <paul.marechal@ericsson.com>