core: fix peer dependency
`@theia/core` depends on `@theia/electron` as an optional peer
dependency. This is because the latter is only useful for Electron
applications, so this really is an optional dependency that should be
installed by application makers as required.

The issue with this is that `lerna` apparently doesn't bump exact peer
dependencies when publishing...

To fix that problem we'll just target `*` (any version) and application
makers will be required to pull the same version of `@theia/electron` as
all others `@theia/xyz` packages.