monaco: properly pass `options` to `input`
The commit updates the `input` method present in `QuickInputService` to
properly pass along `options` so that callers can configure the input to
their needs including placeholders, and input validation.

Previously, the `options` parameter was never passsed internally meaning
the input would not respect the caller's configurations. The code now
properly passes the options to the internal monaco `input` with the
proper types.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>