custom-editor: update opener to respect `option`
The commit updates the `custom-editor` to respect the `option` priority
which is used to control the opener behavior. Plugins may set their
priority to `option` which will not open the custom-editor by default,
and instead default to the regular opener (editor). The custom-editor
should still be a valid opener, when they perform `opens with...`.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>