eslint: add plugin to warn against `src` imports
The commit adds a new `no-src-import` rule to our custom eslint-plugin
in order to warn against the use of `/src/` imports over `/lib/` which
causes the build to fail and is not trivial for developers to identify
as to why. The rule should help write developers more easily identify
when they are using imports incorrectly.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>