problems: add additional tabbar support
The commit adds support for additional decoration types when rendering
tabbars (`fontData` and `tailDecorations`). One use-case added is for
problems to decorate tabs when they have problem markers including the
count.

The changes also align the styling with vscode, where we have
a tail decoration rather than an icon overlay.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>