debug: fix restoration of selected configuration (#10287)
The debug view has the functionality to save the currently selected
debug configuration to local storage. However the restoring of it failed
as the initialization of the debug configuration manager was not
completed by the time the restore action was attempted. The
`initialized` field was assigned after some asynchronous operation
which meant it was `undefined` when other logic tried to use it.

This change fixes the intended wait for initialization of the debug
configuration manger.

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>
Co-authored-by: Paul Marechal <paul.marechal@ericsson.com>