repo: update `publish:latest` script
The commit updates the `publish:latest` script so that `lerna` does not
automatically attempt to push the changes following a latest release
given that the master branch is protected and it would fail. Previously,
due to this error, the publishing had to be manual in order for it not
to fail.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>