ovsx-client: introduce async provider
The commit fixes an issue where previously the backend would never
properly read the cli arguments in time to create the ovsx-client. This
meant that while the api-version was respected when searching in the
view, it was not when installing.

The change removes the use of `OVSXAsyncClient` and instead uses an
`async` provider of `OVSXClient` which will probably be set with the
correct api version and url when ready.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>