tests: skip flaky tests without ignoring them
Add a function to run test cases but skip them on failure in order to
not randomly fail our GitHub action checks.

The flakiness itself should be addressed, but since it is not
straightforward it just keeps failing every PR's CI for no good reason.

Until it can be fixed we shall let it fail silently.