filesystem: fix `createFolder` emitter
The commit fixes an issue where `createFolder` was called by a
user-gesture (through the UI) which did not properly fire the
`onDidRunUserOperation` event. The event is ultimately used by the
plugin-system to properly listen to file/folder creation through the
`vscode.workspace.onDidCreateFiles` event.

The change now aligns with other methods such as `create`, `move` and
`delete`.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>