ci: fix `set-milestone-on-pr` workflow
The commit fixes an issue with the `set-milestone-on-pr` workflow which
did not work for how special a case `@theia/core` is with how it
generates shared exports (and references `@theia/electron`) as a
`peerDependency`. Rather than build the repo before setting the
milestone, we can simply choose a different package.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>