Add missing property to CustomDocumentOpenContext
VSCode defines the optional property `untitledDocumentData` in
`CustomDocumentOpenContext`. In theia.d.ts, this property is however
missing. As it is optional and is not used from the main implementation,
this change is merely achieving source compatibility, but prepares its
usage in the ext implementation of the plugin API.

Contributed on behalf of STMicroelectronics

Signed-off-by: Philip Langer <planger@eclipsesource.com>

Fixes https://github.com/eclipse-theia/theia/issues/10005

Change-Id: I1b4d67b5b11fa2c94a8590aac6b8ad0f040cee94