Keep current configuration when executing a dynamic config.
Starting a session from a dynamic debug configuration left it
selected as `current` in the `DebugConfigurationManager`, however
this selection was not reflected in the debug view as the dynamic
configurations are not found under the `launch.json` file.

Therefore before this change, starting a debug session clicking the
`start` button under the debug view may start a previously run
dynamic configuration instead of the one reflected in the GUI.

This change fixes the above by leaving the current configuration,
and specify the configuration to use in an existing optonal argument.