Multi window support for web views
Support for moving webview-based views into a secondary window/tab.
For webview-based views a new button becomes available in the toolbar of the view to move the view to a secondary window.
This is only support for webview-based views and only one view can be moved into the secondary window.
There can be many secondary windows though.

Contributed on behalf of ST Microelectronics and Ericsson and by ARM and EclipseSource.

Primary changes:
- Add concept of extractable widgets. Only widgets implementing the interface can be extracted
- Add WidgetExtractionHandler that encapsulates logic to move widgets to new windows.
- Only webviews can be extracted
- Configure opened secondary windows in electron
- Bridge plugin communication
- Contribute widget extraction button in a separate extension `widget-extraction-ui`
- Extend application shell areas with a `secondaryWindow` area that contains all extracted widgets
- Extend frontend and webpack generators to generate the base html for secondary windows and copy it to the lib folder

Secure messaging between external webview and Theia:
Webviews only accept messages from its direct parent window. This is necessary to avoid Cross Site Scripting (XSS).
To make the messaging work in secondary windows, messages are sent to the external widget and then delegated to the webview's iframe.
Thereby, the secondary window only accepts messages from its opener which is the theia main window.
To achieve this, a webview knows the secondary window it is in (if any). It then sends messages to this window instead of directly to the webview iframe.