Add correct extension mode to plugin context
The plugin context was hardcoded to have the production extensionMode.
Now a flag is passed to mark a plugin as under development when
deploying it. This flag is stored in the plugin metadata and later set in
the plugin. When creating the context this flag is then converted to the
correct extensionMode.

Fix #10201

Contributed on behalf of STMicroelectronics
Signed-off-by: Eugen Neufeld <eneufeld@eclipsesource.com>