Export page objects directly under `lib` (#11023)
Having `src` in the exported paths of the `lib` folder is unexpected.
Better to export it in a way so that imports for consumers look like
`import { TheiaApp } from '@theia/playwright';`

Therefore, we place define `src`as `rootDir` to get rid of the `src` in
the `lib` path, add an index rexporting all files in the root and define
this index as `main` of this package.

For simplicity we also moved the tests into the `src` folder to be able to
compile them consistently.

This change removes linting from build script to be consistent with the
other packages in Theia. Linting is invoked explicitly now in the CI.
So there is no need anymore to also lint in the yarn build script.