fix quick-open: multi-step input (#8777)
Fixes multiStep quick input flows that call createInputBox.

Previously, when invoking show() a quick input box would be initially
created by calling update({visible: true}) in super(), which in turn
would invoke $createOrUpdate in quick-open-main and create an InputBox
and corresponding session. After the session is created, show() then
calls $showInputBox (An api method introduced by Theia, which is not
supported by the VS Code API) which creates another InputBox, destroying
the previously created session and returning no value to the caller.
This leaves an orpaned InputBox on screen.

To correct this, we simply enable show to invoke $createOrUpdate which
handles the quick input flow correctly.

1. Validate existing multi-step work flows by using the
   [QuickInput](https://github.com/microsoft/vscode-extension-samples/tree/master/quickinput-sample)
   [sample](https://github.com/vince-fugnitto/quick-open-samples/releases/download/1.0.0/quickinput-sample-0.0.1.vsix)

2. Validate quick input works in multi-step flows used by [VS Code
   Solution Explorer](https://open-vsx.org/extension/fernandoescolar/vscode-solution-explorer)
   using the the [reproduction steps provided](https://github.com/eclipse-theia/theia/issues/8777#issuecomment-1045213722)

- [x] As an author, I have thoroughly tested my changes and carefully followed [the review guidelines](https://github.com/eclipse-theia/theia/blob/master/doc/pull-requests.md#requesting-a-review)

- As a reviewer, I agree to review in accordance with [the review guidelines](https://github.com/eclipse-theia/theia/blob/master/doc/pull-requests.md#reviewing)

Signed-off-by: Casey Flynn <caseyflynn@google.com>