lint: fix `runtime-import-check` plugin (#11212)
The commit fixes the `runtime-import-check` plugin which previously did
not work well due to a recent change.

The commit also fixes a `runtime-import-check` error present from
`variable-resolver`. The previous interface `CommandIdVariables` was
defined in `browser` but eventually used in `node`. The commit moves the
interface to `common` so it is safely accessible for `browser` and
`node`.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>