[VSX] Display plugins counts
Adds a plugin count to the title and captions of each instance of
`VSXExtensionsWidget` except for `SEARCH_RESULT` as the count may be
partial.

The solution extends the functionality under view-container by
adding a BadgeWidget interface that can be implemented by widgets
wrapped in a ViewContainerPart, this is the reference implementation
for it.

This solution does not show the badge for view parts that contain a
toolbar populated with more than one item.

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>