Add `Thenable` type to `theia.d.ts`
As discussed in #11214 the `theia.d.ts` should look as close as possible
to the `vscode.d.ts`.
It is currently missing the `Thenable` type and using `PromiseLike` instead.
The goal of this PR is to unify this all to `Thenable`.
Adding the type will also fix and issue in the `vscode-theia-comparator`,
where the Thenable type could not be resolved.

Resolves #11214

Contributed on behalf of STMicroelectronics