file-search: fix test case
The commit fixes the test-case which was not producing results but was passing as a false positive.
The previous implementation was relying on where `theia` was cloned and meant that the test could fail.

The test was updated to harcode the `searchPattern` rather than use the `path` to determine it.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>