vscode: Make TerminalLink compatible to VSCode
Replaces the out-dated definition of the TerminalLink interface with the current class that also provides a constructor.

In the course of fixing this, I discovered #11521.
So it looks like Theia doesn't support registering TerminalLinkProvider
at the moment. Anyway, making TerminalLink compatible to VS Code
will help anyway, but only really gets effective once #11521 is fixed.

Fixes #11507
Contributed on behalf of STMicroelectronics.

Change-Id: If183c73b8152770c4c9062355f92143f7e201e53
Signed-off-by: Philip Langer <planger@eclipsesource.com>