Fix Debug hover might not appear (#11500)
As per the issue title:
`Debug Hover Widget might not appear for evaluated expressions`

Resolving the position of a selection did not resolve if a
`word` was not found within the specific language model.
This caused specific variables provided by the extension to fail,
e.g. `${name}` on a markdown document.

The required matching on the language model was removed,
this is aligned with the solution in `vscode`.

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>