vscode: Support TerminalOptions.hideFromUser
* Add TerminalOptions.hideFromUser in theia.d.ts
* Add TerminalWidget.hiddenFromUser
* Fix cwd for created terminals with undefined cwd (default: workspace)

If TerminalOptions.hideFromUser is set to true, the TerminalWidget
will be hidden from any user-facing terminal lists until it is first
explicitly shown to the user (see TerminalWidget.hiddenFromUser).

As we don't open TerminalWidgets by default on `TerminalService.newTerminal`
and as we don't have a global terminal view (in contrast to VS Code),
we only need to hide it in the `TerminalQuickOpenService` really.

Contributed on behalf of STMicroelectronics.
Fixes https://github.com/eclipse-theia/theia/issues/11144

Change-Id: Ic1fdf2d5515d5b2fa8feb58b72218b93eeabda4d