[CI] Use "force_orphan" option of GH action peaceiris/actions-gh-pages@v3
With this change, the API documentation that we generate as part of our CI
will overwrite the `gh-pages` branch, only keeping the latest generated
documentation. This will reduce the repository size (when cloned, from
close to 3GB to ~150M.

As a consequence, we will lose the history of API documentation. It will
still be possible to manually generate it locally (yarn docs).

Fixes #11625

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>