Support vscode TypeHierarchy API (references-view)
This change introduces support for vscode Typehierarchy API and
integrates the solution with the facilities provided by the
references-view extension.
E.g. via the implementation of the commands
	`vscode.prepareTypeHierarchy`,
	`vscode.provideSupertypes`,
	`vscode.provideSupertypes`

The `Show Type Hierarchy` context menu shall be enabled from an
editor as long as a TypeHIerarchyProvider is registered for the
specific language.

Similarly the following commands shall now be functional from the
command palette.
	`Types: Show Type Hierarchy`
	`Types: Show Subtypes`
	`Types: Show Supertypes`

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>