11685: Restore cancellation token behavior
Ensure that cancelling an RPC call does not result in an automatic rejection on the calling side.
Instead the token is passed to the receiving side and handed there. 
This restores the cancellation strategy that was used with the old json-rpc architecture (prior to 1.28)

Contributed on behalf of STMicroelectronics.
Fixes #11685