Introduce automated license check reviews
In this PR we introduce the option to use our license check tool,
"dash-licenses", in "Automatic IP Team Review Requests" mode [1].
In this mode, any dependency that's found to have an unclear or
suspicious license will be automatically submitted to the Eclipse
Foundation for review. Each such dependency will have a ticket opened
on the Foundation's Gitlab and be automatically reviewed. If the
automated review is not conclusive, a manual assessment will be
performed by the Foundation's IP team.

In our experience, most dependencies are approved automatically
within minutes.

To perform a license check with automated reviews, use the new script:
$> yarn license:check:review

To perform the license check without the automated review, do as before:
$> yarn license:check

Note: for review mode to work, a Personal Access Token from the Foundation's
Gitlab is required, created from a project committer's Gitlab profile.
Set it in an environment variable named "DASH_LICENSES_PAT". E.g. in bash:
$> export DASH_LICENSES_PAT=<token>

[1] https://github.com/eclipse/dash-licenses#automatic-ip-team-review-requests

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>