[CI] Fix license check for non-committers
Be more flexible in the abscense of the EF Gitlab Private Access Token, to
accomodate the case where non-committers (contributors without write access
to the repo) submit PRs and do not have access to GitHub secrets (for
security reasons).

Instead of immediately failing the license check, warn and proceed in
"normal mode" - i.e. like we did before introducing the "review mode".

Fixes #11769

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>