fix: make  on first tab cycle back to last tab
Before this commit, trying to cycle back to the previous tab, when you
were on the first tab had no action. The expected behaviour was to have
the last tab in focus. This was because trying to move back from the first
tab was thought to be an action of moving to the previous tab group. This
commit fixes that issue by treating it as a same tab group action, and
redirecting the user to the last tab in the same group.

Fixes #11746

Signed-off-by: Pranav Joglekar <pranav2000joglekar@gmail.com>