Theia 3PP license check workflow improvement: retry upon failure, after a delay
Currently, if our 3PP license check workflow fails to validate the
license of all our dependencies, the workflow is marked as failed and
the offending dependencies are automatically submitted to the Eclipse
Foundation's Gitlab for analysis and (hopefully) approval. This approval
 can often happen in a totally automated way, within a few minutes.
 Re-running the workflow at that point would end with success, but this
 needs to be triggered manually.

This commit improves the workflow, to have it fail less often.  We do
this by taking advantage of the usually quick approval of 3PP
dependencies that failed the check, by modifying the bash command used
to run the license check:  upon failure, it waits for 15 minutes and
then another attempt is made. Any dependency that was approved in the
meantime will now pass the check.

Before:
$> yarn license:check:review
After:
$> yarn license:check:review || (sleep 15m && yarn license:check:review)
Note that if the workflow works on the first try, no second attempt will
 happen.