[plugin] update signature for activeSignatureHelp
This is done to follow a change in vscode.d.ts, to keep Theia's
version aligned. The change we follow is subtle; changing an
optional property to one that's mandatory but can be undefined:

someProperty?: someType
to
someProperty: someType | undefined

The discrepancy was found with the vscode-theia-comparator tool.
Other similar changes, to properties of basic TS types, were not
detected by the tool, until tweaked locally. These have been
updated as well.

The alignment, for the most part, required no modification of the
implementation code, only a few tweaks here and there.

A couple of affected interfaces, more deeply integrated with
language support, required minor implementation modifications:
SignatureHelpContext and CompletionContext.

Fixes #11828

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>