repo: add `@babel` deps to our baseline
The commit updates our baseline which is used by `dash-licenses` to
report any dependencies which do not satisfy 3pp requirements.

The tool is currently outputting warnings for `@babel` which is still
under review (over 2 weeks). Generally the review is quite fast but
likely due to the holidays there has been no update as of yet.

The commit updates the baseline to include the entries under review
which are being pulled by `@babel/preset-env` which is only a
`devDependency`.

Signed-off-by: vince-fugnitto <vincent.fugnitto@ericsson.com>